Add missing docs and debug implementations
