+++ src/graphics/rectangle.rs
@@ -17,6 +17,10 @@ pub struct Rectangle<T> {
 }
 
 impl Rectangle<f32> {
+    /// Returns true if the given [`Point`] is contained in the [`Rectangle`].
+    ///
+    /// [`Point`]: type.Point.html
+    /// [`Rectangle`]: struct.Rectangle.html
     pub fn contains(&self, point: Point) -> bool {
         self.x <= point.x
             && point.x <= self.x + self.width
