+++ src/input.rs
@@ -1,25 +1,34 @@
 //! Allow players to interact with your game.
 
 mod event;
+mod keyboard_and_mouse;
 
 pub use event::{ButtonState, Event, KeyCode, MouseButton};
-
-use std::collections::HashSet;
+pub use keyboard_and_mouse::KeyboardAndMouse;
 
 use crate::graphics::Point;
 
+/// An input tracker.
 pub trait Input {
+    /// Creates a new [`Input`].
+    ///
+    /// [`Input`]: trait.Input.html
     fn new() -> Self;
 
-    /// Process an input event and keep track of it in your [`Input`] type.
+    /// Processes an input event.
     ///
     /// This function may be called multiple times during event processing,
     /// before [`Game::interact`].
     ///
-    /// [`Input`]: #associatedtype.Input
-    /// [`interact`]: #method.interact
+    /// [`Game::interact`]: ../trait.Game.html#method.interact
     fn update(&mut self, event: Event);
 
+    /// Clears any temporary state that should be consumed by [`Game::interact`]
+    /// and could accumulate otherwise.
+    ///
+    /// This method will be called after each [`Game::interact`].
+    ///
+    /// [`Game::interact`]: ../trait.Game.html#method.interact
     fn clear(&mut self);
 }
 
@@ -33,91 +42,8 @@ impl Input for () {
     fn clear(&mut self) {}
 }
 
+/// A cursor position tracker.
 pub trait HasCursorPosition {
+    /// Returns the current cursor position.
     fn cursor_position(&self) -> Point;
-}
-
-pub struct KeyboardAndMouse {
-    cursor_position: Point,
-    is_cursor_taken: bool,
-    is_mouse_pressed: bool,
-    points_clicked: Vec<Point>,
-    released_keys: HashSet<KeyCode>,
-}
-
-impl KeyboardAndMouse {
-    pub fn cursor_position(&self) -> Point {
-        self.cursor_position
-    }
-
-    pub fn is_cursor_taken(&self) -> bool {
-        self.is_cursor_taken
-    }
-
-    pub fn clicks(&self) -> &[Point] {
-        &self.points_clicked
-    }
-
-    pub fn was_key_released(&self, key_code: &KeyCode) -> bool {
-        self.released_keys.contains(&key_code)
-    }
-}
-
-impl Input for KeyboardAndMouse {
-    fn new() -> KeyboardAndMouse {
-        KeyboardAndMouse {
-            cursor_position: Point::new(0.0, 0.0),
-            is_cursor_taken: false,
-            is_mouse_pressed: false,
-            points_clicked: Vec::new(),
-            released_keys: HashSet::new(),
-        }
-    }
-
-    fn update(&mut self, event: Event) {
-        match event {
-            Event::CursorMoved { x, y } => {
-                self.cursor_position = Point::new(x, y);
-            }
-            Event::CursorTaken => {
-                self.is_cursor_taken = true;
-            }
-            Event::CursorReturned => {
-                self.is_cursor_taken = false;
-            }
-            Event::MouseInput {
-                button: MouseButton::Left,
-                state,
-            } => match state {
-                ButtonState::Pressed => {
-                    self.is_mouse_pressed = !self.is_cursor_taken;
-                }
-                ButtonState::Released => {
-                    if !self.is_cursor_taken && self.is_mouse_pressed {
-                        self.points_clicked.push(self.cursor_position);
-                    }
-
-                    self.is_mouse_pressed = false;
-                }
-            },
-            Event::KeyboardInput {
-                key_code,
-                state: ButtonState::Released,
-            } => {
-                let _ = self.released_keys.insert(key_code);
-            }
-            _ => {}
-        }
-    }
-
-    fn clear(&mut self) {
-        self.points_clicked.clear();
-        self.released_keys.clear();
-    }
-}
-
-impl HasCursorPosition for KeyboardAndMouse {
-    fn cursor_position(&self) -> Point {
-        self.cursor_position
-    }
 }
